pyproject.toml
[project]
name = "touhou"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["Pillow>=10.0.0", "Flask>=3.0.0", "Werkzeug>=3.0.0"]
